convert bdn9 to DD config
