+++ keyboards/keebio/bdn9/bdn9.c
@@ -1,1 +1,46 @@
 #include "bdn9.h"
+
+void eeconfig_init_kb(void) {
+#ifdef RGBLIGHT_ENABLE
+    rgblight_enable();             // Enable RGB by default
+    rgblight_sethsv(0, 255, 255);  // Set default HSV - red hue, full saturation, full brightness
+#    ifdef RGBLIGHT_ANIMATIONS
+    rgblight_mode(RGBLIGHT_MODE_RAINBOW_SWIRL + 2);  // set to RGB_RAINBOW_SWIRL by default
+#    endif
+#endif
+
+#ifdef RGB_MATRIX_ENABLE
+    rgb_matrix_enable();  // Enable RGB by default
+#endif
+
+    eeconfig_update_kb(0);
+    eeconfig_init_user();
+}
+
+#ifdef ENCODER_ENABLE
+bool encoder_update_kb(uint8_t index, bool clockwise) {
+    if (!encoder_update_user(index, clockwise)) { return false; }
+    if (index == 0) {
+        if (clockwise) {
+            tap_code(KC_VOLU);
+        } else {
+            tap_code(KC_VOLD);
+        }
+    }
+    else if (index == 1) {
+        if (clockwise) {
+            tap_code(KC_DOWN);
+        } else {
+            tap_code(KC_UP);
+        }
+    }
+    else if (index == 2) {
+        if (clockwise) {
+            tap_code(KC_PGDN);
+        } else {
+            tap_code(KC_PGUP);
+        }
+    }
+    return false;
+}
+#endif
