+++ keyboards/keebio/bdn9/bdn9.h
@@ -1,7 +1,25 @@
+/* Copyright 2020 Danny Nguyen <danny@keeb.io>
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
 #pragma once
+
+#include "quantum.h"
 
 #ifdef KEYBOARD_keebio_bdn9_rev1
     #include "rev1.h"
-#elif KEYBOARD_keebio_bdn9_rev2
+#endif
+#ifdef KEYBOARD_keebio_bdn9_rev2
     #include "rev2.h"
 #endif
