+++ keyboards/keebio/bdn9/rev1/info.json
@@ -3,5 +3,45 @@
     "usb": {
         "pid": "0x1133",
         "device_version": "1.0.0"
-    }
+    },
+    "processor": "atmega32u4",
+    "bootloader": "caterina",
+    "debounce": 5,
+    "features": {
+        "console": true,
+        "command": true,
+        "extrakey": true,
+        "rgblight": true,
+        "backlight": true,
+    },
+    "matrix_pins": {
+        "direct": [
+            ["D2", "D4", "F4"],
+            ["D7", "B1", "B3"],
+            ["E6", "B4", "B2"],
+        ]
+    },
+    "encoder": {
+        "enabled": true,
+        "rotary": [
+            { "pin_a": "D0", "pin_b": "D1"},
+            { "pin_a": "F6", "pin_b": "F5"},
+            { "pin_a": "B6", "pin_b": "F7"},
+        ]
+    },
+    "backlight": {
+        "pin": "B5",
+        "levels": 7,
+    },
+    "rgblight": {
+        "pin": "D3",
+        "led_count": 4,
+        "sleep": true,
+        "hue_steps": 8,
+        "saturation_steps": 8,
+        "brightness_steps": 8,
+        "animations": {
+            "all": true,
+        }
+    },
 }
